Move function to utils and some minor clenaup.
