+++ ads/google/a4a/utils.js
@@ -960,3 +960,49 @@ export function getAmpRuntimeTypeParameter(win) {
   const art = getBinaryTypeNumericalCode(getBinaryType(win));
   return isCdnProxy(win) && art != '0' ? art : null;
 }
+
+/**
+ * Returns the fixed size of the given element, or the fixed size of its nearest
+ * ancestor that has a fixed size, if the given element has none.
+ * @param {!Element} element
+ * @return {number} The width of the given element, or of the nearest ancestor
+ *    with a fixed size, if the given element has none.
+ */
+export function getContainerWidth(element) {
+    let parent = element;
+    let maxDepth = 100;
+    // Find the first ancestor with a fixed size
+    while (parent && maxDepth--) {
+      const layout = parent.getAttribute('layout');
+      switch (layout) {
+        case Layout.NODISPLAY:
+          return 0;
+        case Layout.FIXED:
+          return parseInt(parent.getAttribute('width'), 10) || 0;
+        case Layout.RESPONSIVE:
+        case Layout.FILL:
+        case Layout.FIXED_HEIGHT:
+        case Layout.FLUID:
+          // The above layouts determine the width of the element by the
+          // containing element, or by CSS max-width property.
+          if (parent.style.maxWdith) {
+            return parseInt(parent.style.maxWidth, 10) || 0;
+          }
+          parent = parent.parentElement;
+        case Layout.CONTAINER:
+          // Container layout allows the container's size to be determined by
+          // the children within it, so in principle we can grow as large as the
+          // viewport.
+          const viewport = Services.viewportForDoc(element);
+          return viewport.getSize().width;
+        case Layout.FLEX_ITEM:
+          // Width is determined in part by sibling elements. This is usually
+          // used for carousel-type constructs, and is unreasonably difficult to
+          // compute the actual available width.
+        default:
+          return 0;
+      }
+    }
+    return 0;
+  }
+}
