+++ extensions/amp-ad-network-doubleclick-impl/0.1/amp-ad-network-doubleclick-impl.js
@@ -34,6 +34,7 @@ import {
   ValidAdContainerTypes,
   addCsiSignalsToAmpAnalyticsConfig,
   extractAmpAnalyticsConfig,
+  getContainerWidth,
   getCsiAmpAnalyticsConfig,
   getCsiAmpAnalyticsVariables,
   getEnclosingContainerTypes,
@@ -93,7 +94,6 @@ import {
   lineDelimitedStreamer,
   metaJsonCreativeGrouper,
 } from '../../../ads/google/a4a/line-delimited-response-handler';
-import {modifyIfHeightNotExpandable} from './flexible-slot-utils';
 import {parseQueryString} from '../../../src/url';
 import {setStyles} from '../../../src/style';
 import {stringHash32} from '../../../src/string';
@@ -506,9 +506,9 @@ export class AmpAdNetworkDoubleclickImpl extends AmpA4A {
       'fluid': this.isFluidRequest_ ? 'height' : null,
       'fsf': this.forceSafeframe ? '1' : null,
       'msz': this.sendFlexibleAdSlotParams ?
-          this.getContainerSizeParameter(this.element) : null,
+          `${getContainerWidth(this.element)}x-1` : null,
       'psz': this.sendFlexibleAdSlotParams ?
-          this.getContainerSizeParameter(this.element.parentElement) : null,
+          `${getContainerWidth(this.element.parentElement)}x-1` : null,
       'scp': serializeTargeting(
           (this.jsonTargeting && this.jsonTargeting['targeting']) || null,
           (this.jsonTargeting &&
@@ -1411,47 +1411,6 @@ export class AmpAdNetworkDoubleclickImpl extends AmpA4A {
   isFluidRequest() {
     return this.isFluidRequest_;
   }
-
-  getContainerSizeParameter(element) {
-    let parent = element;
-    let maxDepth = 100;
-    // Find the first ancestor with a fixed size
-    while (parent && maxDepth--) {
-      const layout = parent.getAttribute('layout');
-      switch (layout) {
-        case Layout.NODISPLAY:
-          return null;
-        case Layout.FIXED:
-          const width = Number(parent.getAttribute('width'));
-          return `${parentWidth}x-1`;
-        case Layout.RESPONSIVE:
-        case Layout.FILL:
-        case Layout.FIXED_HEIGHT:
-        case Layout.FLUID:
-          // The above layouts determine the width of the element by the
-          // containing element, or by CSS max-width property.
-          if (parent.style.maxWdith) {
-            return `${parent.style.maxWidth}x-1`;
-          }
-          parent = parent.parentElement;
-        case Layout.CONTAINER:
-          // Container layout allows the container's size to be determined by
-          // the children within it, so in principle we can grow as large as the
-          // viewport.
-          const viewport = Services.viewportForDoc(this.getAmpDoc());
-          const width = viewport.getSize().width;
-          return `${width}x-1`;
-        case Layout.FLEX_ITEM:
-          // Width is determined in part by sibling elements. This is usually
-          // used for carousel-type constructs, and is unreasonably difficult to
-          // compute the actual available width.
-        default:
-          return null;
-      }
-    }
-    return null;
-  }
-}
 
 AMP.extension(TAG, '0.1', AMP => {
   AMP.registerElement(TAG, AmpAdNetworkDoubleclickImpl);
